ICU update to 54.1 step 7
1. Fix a Windows build failure due to:
   a. 'signed vs unsigned' comparison
   b. 'possible data loss' in conversion : Apply pkasting's patch at
      http://bugs.icu-project.org/trac/ticket/11104

2. Drop a few currencies to cut down the data size by 50kB for non-Android
   platforms.

2. Build the ICU data for Android and check in.
  - Drop all display names for languages/scripts/regions except for zh-Han{s,t}
    as before. ( ~ 1.2MB reduction)
  - Drop cjdict by applying android/brkitr.patch. (~ 2MB reduction)
  - Include the display names for only 60+ currencies ( ~ 400kB reduction
    from the non-Android data.
  - Minimize the locale data for 9 locales Chrome on Android is not localized
    to. Drop currency names for those 9 locales. ( ~ 150kB reduction)

Size change:
  1. Non-android: 10,255,584 to 10,200,880
  2. Android:
     - Final : 6,270,880
       With 60+ currency names added (for bug 370849) and
       9 unnecessary locale data dropped.
       It's 232,240 bytes larger than ICU 52.1 (6,038,640).
     - Without any currency names but with 9 unnecessary locale data: 6,026,816
     - With 60+ currency names and 9 unnecessary locale data: 6,426,368

BUG=370849,428145
TEST=Build on Windows. Blink layout tests, webkit unittests.
R=mark@chromium.org, wangxianzhu@chromium.org

Review URL: https://codereview.chromium.org/877193003

+++ source/i18n/digitlst.cpp
@@ -704,7 +704,7 @@ DigitList::set(int64_t source)
     U_ASSERT(uprv_strlen(str) < sizeof(str));
 
     uprv_decNumberFromString(fDecNumber, str, &fContext);
-    internalSetDouble(source);
+    internalSetDouble(static_cast<double>(source));
 }
 
 /**
